+++ examples/data-sources/woodpecker_organization_secret/data-source.tf
@@ -1,0 +1,4 @@
+data "woodpecker_organization_secret" "secret" {
+  repo_owner = "example_org"
+  name       = "example secret"
+}