+++ examples/resources/woodpecker_organization_secret/import.sh
@@ -1,0 +1,2 @@
+# Syntax: <owner>/<name>
+terraform import woodpecker_organization_secret.secret "example_org/test secret"