+++ examples/resources/woodpecker_organization_secret/resource.tf
@@ -1,0 +1,5 @@
+resource "woodpecker_organization_secret" "secret" {
+  owner = "example_org"
+  name  = "example secret"
+  value = "example value"
+}